+++ spec/integration/mongoid/mongoid_spec.rb
@@ -26,6 +26,7 @@ describe 'Mongoid' do
   it_behaves_like 'by calendar month'
   it_behaves_like 'by quarter'
   it_behaves_like 'by week'
+  it_behaves_like 'by cweek'
   it_behaves_like 'by weekend'
   it_behaves_like 'by year'
   it_behaves_like 'relative'
@@ -33,14 +34,29 @@ describe 'Mongoid' do
   it_behaves_like 'scope parameter'
 
   describe '#between_times' do
-    subject { Post.between_times(Time.parse('2014-01-01'), Time.parse('2014-01-06')) }
+    subject { Post.between_times(Time.zone.parse('2014-01-01'), Time.zone.parse('2014-01-06')) }
     it { should be_a(Mongoid::Criteria) }
     it { expect(subject.count).to eql(3) }
+
+    context ':order option' do
+
+      it 'should be able to order the result set asc' do
+        scope = Post.by_year(Time.zone.now.year, :order => :asc)
+        expect(scope.options[:sort]).to eq({'created_at' => 1})
+        expect(scope.first.created_at).to eq Time.zone.parse('2014-01-01 17:00:00')
+      end
+
+      it 'should be able to order the result set desc' do
+        scope = Post.by_year(Time.zone.now.year, :order => :desc)
+        expect(scope.options[:sort]).to eq({'created_at' => -1})
+        expect(scope.first.created_at).to eq Time.zone.parse('2014-04-15 17:00:00')
+      end
+    end
   end
 
   describe '#between' do
     it 'should not override Mongoid between method' do
-      posts = Post.between(created_at: Time.parse('2014-01-01')..Time.parse('2014-01-06'))
+      posts = Post.between(created_at: Time.zone.parse('2014-01-01')..Time.zone.parse('2014-01-06'))
       expect(posts.count).to eql(3)
     end
   end